Moved Display code to GenerateGraphs
